+++ bin/commands/about.js
