+++ bin/commands/add.js
