+++ bin/commands/all.js
