+++ bin/commands/help.js
