+++ bin/commands/score.js
