+++ bin/commands/today.js
