+++ bin/regex.js
