+++ bin/storage.js
